give server genesis block at startup
