+++ modules/host/announce_test.go
@@ -18,6 +18,7 @@ func TestAnnouncement(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
+	ht.tpUpdateWait()
 
 	// Check that the announcement made it into the transaction pool correctly.
 	txns := ht.tpool.TransactionSet()
@@ -30,4 +31,7 @@ func TestAnnouncement(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
+
+	// TODO: Need to check that the host announcement gets the host into the
+	// hostdb.
 }
